+++ pkg/runner/step.go
@@ -5,17 +5,23 @@ import (
 	"testing"
 
 	"github.com/kyverno/chainsaw/pkg/apis/v1alpha1"
-	"github.com/kyverno/chainsaw/pkg/client"
 	"github.com/kyverno/chainsaw/pkg/resource"
+	runnerclient "github.com/kyverno/chainsaw/pkg/runner/client"
 	"github.com/kyverno/chainsaw/pkg/runner/logging"
-	"github.com/kyverno/chainsaw/pkg/runner/operations"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	ctrlclient "sigs.k8s.io/controller-runtime/pkg/client"
 )
 
-func executeStep(t *testing.T, logger logging.Logger, ctx Context, basePath string, config v1alpha1.ConfigurationSpec, test v1alpha1.TestSpec, step v1alpha1.TestSpecStep) {
+func executeStep(
+	t *testing.T,
+	logger logging.Logger,
+	client runnerclient.Client,
+	ctx Context,
+	basePath string,
+	config v1alpha1.ConfigurationSpec,
+	test v1alpha1.TestSpec,
+	step v1alpha1.TestSpecStep,
+) {
 	t.Helper()
-	c := ctx.clientFactory(logger)
 	stepCtx, cancel := timeoutCtx(config, test, step.Spec)
 	defer cancel()
 	for _, delete := range step.Spec.Delete {
@@ -29,24 +35,24 @@ func executeStep(t *testing.T, logger logging.Logger, ctx Context, basePath stri
 			logger.Log(err)
 			t.FailNow()
 		}
-		if err := operations.Delete(stepCtx, logger, &resource, c); err != nil {
+		if err := client.Delete(stepCtx, &resource); err != nil {
 			logger.Log(err)
 			t.FailNow()
 		}
 	}
-	var cleanup operations.CleanupFunc
-	if skip := skipDelete(config, test, step.Spec); skip == nil || !*skip {
-		cleanup = func(obj ctrlclient.Object, c client.Client) {
-			t.Cleanup(func() {
-				cleanupCtx, cancel := timeoutCtx(config, test, step.Spec)
-				defer cancel()
-				if err := operations.Delete(cleanupCtx, logger, obj, c); err != nil {
-					logger.Log(err)
-					t.FailNow()
-				}
-			})
-		}
-	}
+	// var cleanup operations.CleanupFunc
+	// if skip := skipDelete(config, test, step.Spec); skip == nil || !*skip {
+	// 	cleanup = func(obj ctrlclient.Object, c client.Client) {
+	// 		t.Cleanup(func() {
+	// 			cleanupCtx, cancel := timeoutCtx(config, test, step.Spec)
+	// 			defer cancel()
+	// 			if err := client.Delete(cleanupCtx, logger, obj, c); err != nil {
+	// 				logger.Log(err)
+	// 				t.FailNow()
+	// 			}
+	// 		})
+	// 	}
+	// }
 	for _, apply := range step.Spec.Apply {
 		resources, err := resource.Load(filepath.Join(basePath, apply.File))
 		if err != nil {
@@ -59,7 +65,7 @@ func executeStep(t *testing.T, logger logging.Logger, ctx Context, basePath stri
 				logger.Log(err)
 				t.FailNow()
 			}
-			if err := operations.Apply(stepCtx, logger, resource, c, cleanup); err != nil {
+			if err := client.Apply(stepCtx, resource); err != nil {
 				logger.Log(err)
 				t.FailNow()
 			}
@@ -77,7 +83,7 @@ func executeStep(t *testing.T, logger logging.Logger, ctx Context, basePath stri
 				logger.Log(err)
 				t.FailNow()
 			}
-			if err := operations.Assert(stepCtx, logger, resources[i], c); err != nil {
+			if err := client.Assert(stepCtx, resources[i]); err != nil {
 				logger.Log(err)
 				t.FailNow()
 			}
@@ -95,7 +101,7 @@ func executeStep(t *testing.T, logger logging.Logger, ctx Context, basePath stri
 				logger.Log(err)
 				t.FailNow()
 			}
-			if err := operations.Error(stepCtx, logger, resources[i], c); err != nil {
+			if err := client.Error(stepCtx, resources[i]); err != nil {
 				logger.Log(err)
 				t.FailNow()
 			}
